Move event_iterator and Analyses
- This commit moves the `event_iterator` module and `Analyses` structure currently located directly in `main.rs` to new `analyses` module
- Moves most of the code from run_analysis to member methods of `Analyses` struct
