+++ src/analyses/analysis/callback_duration.rs
