+++ src/analyses/analysis/publication_in_callback.rs
