+++ src/analyses/analysis/spin_duration.rs
